requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# PDF Processing Libraries
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
tabula-py==2.8.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Database Connectivity
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.4.7

# Azure Storage (NEW)
azure-storage-blob==12.19.0

# Additional Dependencies for PDF processing
opencv-python==4.8.1.78
Pillow==10.1.0

# HTTP and Utilities
requests==2.31.0
urllib3==2.1.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Logging and Configuration
python-dotenv==1.0.0

# WSGI Server for Production
gunicorn==21.2.0